net/hns3: decrease the count when secondary process exits
The "secondary_cnt" will be increased when a secondary process initialized.
But the value of this variable is not decreased when the secondary process
exits, which causes the primary process senses that the secondary process
still exists. As a result, the primary process fails to send messages to
the secondary process after the secondary process exits.

Fixes: 23d4b61fee5d ("net/hns3: support multiple process")
Cc: stable@dpdk.org

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>