hash: fix thash gfni implementation
This patch replaces _mm512_set_epi8 with _mm512_set_epi32
due to the lack of support by some compilers.
Also this patch checks if AVX512F is supported along with GFNI.
This is done if the code is built on a platform that supports GFNI,
but does not support AVX512.

Fixes: 4fd8c4cb0de1 ("hash: add new Toeplitz hash implementation")
Cc: vladimir.medvedkin@intel.com

Signed-off-by: Vladimir Medvedkin <vladimir.medvedkin@intel.com>
Acked-by: Lance Richardson <lance.richardson@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>