test/bpf_convert: do not require DNS lookup
These tests are testing with strings with hostname (dpdk.org)
and this makes test code do a DNS lookup. In container test
runs, DNS is unavailable.  Replace dpdk.org with the current
IP and IPV6 addresses.  Actual addresses don't matter the purpose
of the test is to BPF code conversion, not any packets.

Reported-by: Luca Boccassi <bluca@debian.org>
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>