eal: implement thread priority management functions
Add functions for setting and getting the priority of a thread.
Priorities on multiple platforms are similarly determined by
a priority value and a priority class/policy.

Currently in DPDK most threads operate at the OS-default
priority level but there are cases when increasing the
priority is useful. For example, high performance applications
may require elevated priority levels.

For these reasons, EAL will expose two priority levels which
are named suggestively "normal" and "realtime_critical"
and are computed as follows:

On Linux, the following mapping is created:
RTE_THREAD_PRIORITY_NORMAL corresponds to
* policy SCHED_OTHER
* priority value:   (sched_get_priority_min(SCHED_OTHER) +
                     sched_get_priority_max(SCHED_OTHER))/2;
RTE_THREAD_PRIORITY_REALTIME_CRITICAL corresponds to
* policy SCHED_RR
* priority value: sched_get_priority_max(SCHED_RR);

On Windows, the following mapping is created:
RTE_THREAD_PRIORITY_NORMAL corresponds to
* class NORMAL_PRIORITY_CLASS
* priority THREAD_PRIORITY_NORMAL
RTE_THREAD_PRIORITY_REALTIME_CRITICAL corresponds to
* class REALTIME_PRIORITY_CLASS
* priority THREAD_PRIORITY_TIME_CRITICAL

Note that on Linux the resulting priority value will be 0,
in accordance to the docs that mention the value
should be 0 for SCHED_OTHER policy.

Signed-off-by: Narcisa Vasile <navasile@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>