vhost: fix get hpa fail from guest pages
When processing front-end memory regions messages,
vhost saves the guest/host physical address mappings to guest pages
and merges adjacent contiguous pages if hpa is contiguous,
however gpa is likely not contiguous in PA mode
and merging will cause the gpa range to change.
This patch distinguishes the case of discontinuous gpa
and does a range lookup on gpa when doing a binary search.

Fixes: e246896178e("vhost: get guest/host physical address mappings")
Fixes: 6563cf92380 ("vhost: fix async copy on multi-page buffers")

Signed-off-by: Yuan Wang <yuanx.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>