devtools: spell check
A spell check script that checks for spelling errors in modified .rst files
using git. There is also an optional flag to have the script check the entire
doc directory for errors. The dictionary is used for word exceptions, and the
regex file will exclude any regex patterns supplied on each line.

Errors will be logged and printed to console, and by default the log will be
stored in devtools. The output directory can be modified with a flag.

The script can be executed from anywhere within the dpdk repo.

Signed-off-by: Henry Nadeau <hnadeau@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>