app/testpmd: add GPU memory option for mbuf pools
This patch introduces GPU memory in testpmd through the gpudev library.
Testpmd can be used for network benchmarks when using GPU memory
instead of regular CPU memory to send and receive packets.
This option is currently limited to iofwd engine to ensure
no workload is applied on packets not accessible from the CPU.

The options chose is --mbuf-size so buffer split feature across
different mempools can be enabled.

Signed-off-by: Elena Agostini <eagostini@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>