doc: add build_root to STRIP_FROM_PATH
examples.dox is built inside builddir and so doxygen generates some
dir_HASH.html that includes the builddir name and this may prevent DPDK
documentation to be correctly generated in some distributions, for
example CentOS Stream 9 and RHEL9, since the builddir includes the
architecture.

This commit adds builddir to STRIP_FROM_PATH in order to avoid
generating html files that includes it.

Fixes: a6090630f4e5 ("doc: automate examples file list for API")
Cc: ferruh.yigit@intel.com
Signed-off-by: Timothy Redaelli <tredaelli@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>