net/mlx5: fix RSS validation for meter hierarchy
In a meter hierarchy, all the meters are marked with having RSS if
the final meter's termination action is RSS.

When validating a flow rule with meter hierarchy, the RSS action
should not be fetched from the current meter if it is not the final
one.

The fate action union is next meter ID instead of the pointer to the
RSS action. By using the final meter in the hierarchy, the flow rule
validation will succeed without any crash caused by the invalid RSS
action pointer access.

Fixes: 1ce19ab1f43b ("net/mlx5: fix RSS validation with meter policy")
Cc: stable@dpdk.org

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Reviewed-by: Li Zhang <lizh@nvidia.com>
Reviewed-by: Shun Hao <shunh@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>